+++ CHANGELOG.md
@@ -3,4 +3,13 @@
 * Initial release support Quickbooks Payments, Quickbooks Accounting and Authorization.
 
 ## 0.0.2
+
 * Update documentation for clients' methods.
+
+## 0.0.3
+
+* Added copyWith to Customer and BankAccount models.
+
+## 0.0.4 
+
+* Added helper function setCachedRealmId to set the realmId used in the client.